Update plugin.py
Add extra device called 'Available". This Domoticz device shows whether the Google device is still available... This is an alternative information for the 'Timeout' function. I am using the plugin to detect also if my television is ON (Chromecast available) or not. However I want to avoid the red timeout bars.